Device bootloader: add oscillator calibration support.
A new variable in flash, storing the value for the oscillator tuning register has been added.
A new procedure has been added, which attempts to read this variable and set the register. If the variable is not set, will go into a calibration procedure, which requires a connection to a USB host. Once successful, the value will be stored. The stat LED will blink quickly while waiting for a host connection.

Upon reset:
- If "boot" pin not grounded, run the new tuning procedure and then the app.
- Otherwise:
  - If pin 1 is grounded - erase the configuration page (including the osctune variable).
  - Run the new tuning procedure.
  - Go to bootloader mode.

This means that the behavior of a freshly-programmed IOIO would be:
- If powered and not connected to a host, the LED will blink quickly and the IOIO won't work.
- As soon as connected to a host, the app will run.

Upon next reset, the app would run immediately.