firmware: Upgrade btstack to r2027
The main purpose of this upgrade was to add the CoD report, with the hope
that some Android devices which reject the CoD of 0 (misc.) will now see
the IOIO.
As a side-effect, this version of btstack supports some new features, and
mainly SSP (secure simple pairing).
Since the IOIO doesn't have either a display nor a keyboard, the only
possible pairing option is "Just Works", which lacks authentication. I
opted to disable SSP for the time being, but it is not easy to turn on.