bootloaders / BLAPI: Bug in address map
It turns out that there was a bug in both the bootloader and the app linker
scripts, which allowed the app image to extend to the bootloader settings
page, and hence erase the settings when programmed.
This would only have been felt if the app firmware actually used this page,
which until now hasn't happened.
While this could be considered a BLAPI change, I'm considering it an
implementation bug and not updating the BLAPI version number.
After this fix, both the app linker script now enforces that the app
doesn't extend into the last page. A follow-up commit fixes the
bootloaders.