IOIOLibAccessory: Rewrite Accesory connection for increased reliability
There has been a lot of flakiness in the way the AOA connection was
handled, mostly due to many things in the connection process that can go
wrong, which results in a fairly complex state machine.

To avoid this complexity, the entire connection process has been
collapsed to a single method which attempts the entire connection
process as one atomic linear sequence and the waitForConnect() process
simply involves invoking it repeatedly until it succeeds. While being
slightly less efficient, it is more obviously correct and appears to
address the unreliability issues.