device_bootloader: Fix mysterious bug!
Some boards were observed to not perform the oscillator calibration successfully.
It turns out, that in some cases those boards have managed to go into USB POWERED_STATE and
advance past it in the same USBTasks() call, thus never go past the fast blink stage.

Greetings to my man in SparkFun for the relentless help in getting to the bottom of this
nasty one!