firmware: Small bugfix in UART
Previously, if there was a parity or framing error, we would erroneuosly
discard the *previous* byte if exists keep the corrupt one. This is
because reading RXREG shifts the FERR and PERR bits along with the data.
This fix, reading the error bits first, then shifts out the data and
considers the error.